Fix HPC-X library path prefixes
The HPC-X distribution contains several .pc, .la, and .cmake files with
inaccurate hardcoded prefixes for library file locations, making it
impossible to link with certain libraries.

For example, trying to link against UCX in a CMake build would redirect
the linker to a nonexistent path due to its ucx-targets.cmake file,
and then fail, even with a correct LIBRARY_PATH environment variable.

This commit adds code to update all detected hardcoded prefixes to the
HPC-X path used in this image with a find-and-replace operation
during installation.