Fix error in the prefix workaround when there is nothing to fix
If the grep search for files with invalid hardcoded paths
fails to find any matches, sed throws an error by default
on account of being given no input files to operate on.
This adds flags to xargs to avoid invoking sed at all when
no malformed files are found, as well as to set an appropriate delimiter
to match grep's output.